pyproject.toml
[project]
name = "navimoderator-auto-clip"
version = "0.0.1"
description = "Add your description here"
readme = "README.md"
authors = [
    {name="Gengaru", email="gengaru617science@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "langchain-community>=0.2.15",
    "langchain>=0.2.15",
    "langgraph>=0.2.56",
    "openai>=1.35.13",
    "pandas>=2.2.3",
    "pydantic>=2.9.2",
    "pypdf>=4.3.1",
    "tomli-w>=1.0.0",
    "tomli>=2.0.1",
    "jinja2>=3.1.4",
    "litellm>=1.52.16",
    "setuptools>=75.6.0",
    "coverage>=7.6.10",
    "pytest-cov>=6.0.0",
    "pytest-asyncio>=0.24.0",
    "requests>=2.31.0",
    "fastapi>=0.115.5",
    "google-api-python-client>=2.156.0",
    "freezegun>=1.5.1",
    "pytube>=15.0.0",
    "pytest-mock>=3.14.0",
]

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "pre-commit>=3.8.0",
    "pytest>=8.3.3",
    "setuptools>=75.6.0",
    "twine>=6.0.1",
]

[tool.ruff.lint]
ignore = ["E731","E741","F841","F821"]

[tool.setuptools.packages.find]
where = ["src"]
include = ["navimoderator-auto-clip*"]
